ksupport: refactor build codegen into separate crate
