+++ artiq/firmware/ksupport/Cargo.toml
@@ -11,13 +11,7 @@ crate-type = ["staticlib"]
 
 [build-dependencies]
 build_misoc = { path = "../libbuild_misoc" }
-ddb_parser = { path = "../ddb_parser" }
-itertools = "0.12"
-prettyplease = "0.2.16"
-proc-macro2 = "=1.0.86"  # must match a patched version working with rustc-1.57-nightly
-quote = "1"
-regex = "1"
-syn = "2"
+build_ksupport = { path = "../libbuild_ksupport" }
 
 [dependencies]
 cslice = { version = "0.3" }
@@ -32,3 +26,4 @@ libc = { path = "../libc" }
 unwind = { path = "../libunwind" }
 num_enum = { version = "0.7.2", default-features = false }
 bitflags = "2.4"
+sinara_config = { path = "../libsinara_config" }
