+++ artiq/firmware/ksupport/sinara/urukul/config.rs
@@ -1,4 +1,5 @@
 use core::convert::{TryFrom, TryInto};
+use sinara_config::urukul::{ClkDiv, ClkSel, SyncSel};
 
 /// Urukul configuration register.
 #[derive(Debug, Clone, Eq, PartialEq)]
@@ -12,78 +13,6 @@ pub struct Config {
     pub reset: bool,
     pub io_reset: bool,
     pub clk_div: ClkDiv,
-}
-
-/// Input clock divider.
-#[derive(Debug, Clone, Copy, Eq, PartialEq, IntoPrimitive, TryFromPrimitive)]
-#[repr(u8)]
-pub enum ClkDiv {
-    Default = 0,
-
-    /// Divide by one.
-    One = 1,
-
-    /// Divide by two.
-    Two = 2,
-
-    /// Divide by four.
-    Four = 3,
-}
-
-impl Default for ClkDiv {
-    fn default() -> Self {
-        Self::Default
-    }
-}
-
-impl ClkDiv {
-    /// Reference clock divider, as integer.
-    pub fn divider(&self) -> i32 {
-        match self {
-            Self::Default | Self::Four => 4,
-            Self::One => 1,
-            Self::Two => 2,
-        }
-    }
-}
-
-/// Reference clock input selection.
-///
-/// Only supports Urukul v1.3+.
-#[derive(Debug, Clone, Copy, Eq, PartialEq, IntoPrimitive, TryFromPrimitive)]
-#[repr(u8)]
-pub enum ClkSel {
-    /// On-board crystal, 100 MHz.
-    Internal = 0,
-
-    /// Front-panel SMA connector.
-    Sma = 1,
-
-    /// Internal MMCX connector.
-    Mmcx = 2,
-}
-
-impl Default for ClkSel {
-    fn default() -> Self {
-        Self::Internal
-    }
-}
-
-/// DDS inter-chip synchronization clock source selection.
-#[derive(Debug, Clone, Copy, Eq, PartialEq, IntoPrimitive, TryFromPrimitive)]
-#[repr(u8)]
-pub enum SyncSel {
-    /// EEM1 connector, LVDS pair 0.
-    Eem = 0,
-
-    /// On-board DDS chip 0.
-    Dds0 = 1,
-}
-
-impl Default for SyncSel {
-    fn default() -> Self {
-        Self::Dds0
-    }
 }
 
 impl Config {
