+++ artiq/firmware/libsinara_config/src/lib.rs
@@ -1,0 +1,3 @@
+#![no_std]
+
+pub mod urukul;
